fix Actualizar tipos de ID en CineRepository de Integer a Long y corregir la URL de la base de datos en application.properties

+++ src/main/java/com/example/CineSpringBoot/repositories/CineRepository.java
@@ -4,5 +4,5 @@ import com.example.CineSpringBoot.entities.Cine;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CineRepository extends BaseRepository<Cine, Integer> {
+public interface CineRepository extends BaseRepository<Cine, Long> {
 }
